+++ cluster/deposit.go
@@ -2,18 +2,6 @@
 
 package cluster
 
-import (
-	eth2p0 "github.com/attestantio/go-eth2-client/spec/phase0"
-
-	"github.com/obolnetwork/charon/app/errors"
-	"github.com/obolnetwork/charon/app/z"
-)
-
-const (
-	minSingleDepositAmountGwei = 1000000000  // 1ETH
-	maxTotalDepositAmountGwei  = 32000000000 // 32ETH
-)
-
 // DepositData defines the deposit data to activate a validator.
 // https://github.com/ethereum/consensus-specs/blob/dev/specs/phase0/beacon-chain.md#depositdata
 type DepositData struct {
@@ -56,41 +44,4 @@ func depositDataFromJSON(d depositDataJSON) DepositData {
 		Amount:                d.Amount,
 		Signature:             d.Signature,
 	}
-}
-
-// VerifyDepositAmounts verifies various conditions about partial deposits rules.
-func VerifyDepositAmounts(amounts []eth2p0.Gwei) error {
-	if len(amounts) == 0 {
-		return nil
-	}
-
-	var sum eth2p0.Gwei
-	for _, amount := range amounts {
-		if amount < minSingleDepositAmountGwei {
-			return errors.New("each partial deposit amount must be greater than 1ETH", z.U64("amount", uint64(amount)))
-		}
-
-		sum += amount
-	}
-
-	if sum > eth2p0.Gwei(maxTotalDepositAmountGwei) {
-		return errors.New("sum of partial deposit amounts must sum up to 32ETH", z.U64("sum", uint64(sum)))
-	}
-
-	return nil
-}
-
-// IntsToGweis converts amounts from []int to []eth2p0.Gwei.
-// For verification, please see VerifyDepositAmounts().
-func IntsToGweis(amounts []int) []eth2p0.Gwei {
-	if amounts == nil {
-		return nil
-	}
-
-	var gweiAmounts []eth2p0.Gwei
-	for _, amount := range amounts {
-		gweiAmounts = append(gweiAmounts, eth2p0.Gwei(amount))
-	}
-
-	return gweiAmounts
 }
