+++ dkg/dkg.go
@@ -116,9 +116,9 @@ func Run(ctx context.Context, conf Config) (err error) {
 	}
 
 	if len(conf.DepositAmounts) > 0 {
-		amounts := cluster.IntsToGweis(conf.DepositAmounts)
+		amounts := deposit.IntsToGweis(conf.DepositAmounts)
 
-		if err := cluster.VerifyDepositAmounts(amounts); err != nil {
+		if err := deposit.VerifyDepositAmounts(amounts); err != nil {
 			return err
 		}
