+++ eth2util/deposit/deposit.go
@@ -18,6 +18,11 @@ import (
 	"github.com/obolnetwork/charon/tbls"
 )
 
+const (
+	minSingleDepositAmountGwei = 1000000000  // 1ETH
+	maxTotalDepositAmountGwei  = 32000000000 // 32ETH
+)
+
 var (
 	// https://github.com/ethereum/consensus-specs/blob/dev/specs/phase0/validator.md#eth1_address_withdrawal_prefix
 	eth1AddressWithdrawalPrefix = []byte{0x01}
@@ -185,4 +190,41 @@ type depositDataJSON struct {
 	ForkVersion           string `json:"fork_version"`
 	NetworkName           string `json:"network_name"`
 	DepositCliVersion     string `json:"deposit_cli_version"`
+}
+
+// VerifyDepositAmounts verifies various conditions about partial deposits rules.
+func VerifyDepositAmounts(amounts []eth2p0.Gwei) error {
+	if len(amounts) == 0 {
+		return nil
+	}
+
+	var sum eth2p0.Gwei
+	for _, amount := range amounts {
+		if amount < minSingleDepositAmountGwei {
+			return errors.New("each partial deposit amount must be greater than 1ETH", z.U64("amount", uint64(amount)))
+		}
+
+		sum += amount
+	}
+
+	if sum > eth2p0.Gwei(maxTotalDepositAmountGwei) {
+		return errors.New("sum of partial deposit amounts must sum up to 32ETH", z.U64("sum", uint64(sum)))
+	}
+
+	return nil
+}
+
+// IntsToGweis converts amounts from []int to []eth2p0.Gwei.
+// For verification, please see VerifyDepositAmounts().
+func IntsToGweis(amounts []int) []eth2p0.Gwei {
+	if amounts == nil {
+		return nil
+	}
+
+	var gweiAmounts []eth2p0.Gwei
+	for _, amount := range amounts {
+		gweiAmounts = append(gweiAmounts, eth2p0.Gwei(amount))
+	}
+
+	return gweiAmounts
 }
